fix: critical bug fixes, financial precision, and deployment readiness
- Fix expense logging: add missing participants relation to select queries
- Settlement model: double -> Decimal for OMR precision (3 decimal places)
- Firebase: add initialization in main.dart (FCM was crashing)
- iOS: add camera/photo permission strings, fix deployment target to 15.5
- Thawani: move hardcoded API keys to String.fromEnvironment
- Receipt service: fix column name mismatch (receipt_path -> receipt_url)
- Local database: fix race condition with Completer pattern
- Settings service: add bounds checking for corrupted preferences
- Connectivity provider: add mounted check after async operation
- Gear screen: fix context.pop -> Navigator.pop, claim-then-pack race condition
- Vault screen: remove broken OpenFilex.open with remote URL
- Memories screen: add mounted check after async upload
- Activity service: fix map mutation (create copy before modifying)
- Sync service: fix table name and connectivity check method
- Ledger: Decimal division type fix, add amount validation
- Remove unused deps: riverpod_annotation, flutter_svg, google_mlkit_text_recognition
- Add package_info_plus dependency (was imported but missing)
- Add data/run-logs/ to .gitignore
- Add offline-first design doc and name-based members implementation plans

+++ lib/core/services/cache_service.dart
@@ -163,7 +163,7 @@ class CacheService {
         tripId: map['trip_id'] as String,
         payerParticipantId: map['payer_participant_id'] as String?,
         recipientParticipantId: map['recipient_participant_id'] as String?,
-        amount: double.parse(map['amount'] as String),
+        amount: Decimal.parse(map['amount'] as String),
         note: map['note'] as String?,
         settledAt: DateTime.parse(map['created_at'] as String),
         isDeleted: (map['is_deleted'] as int?) == 1,
